libnss_tcb: Match interfaces with NSS documentation.
According to [1] the interfaces provided are of return-type 'enum nss_status',
of which the _nss_database_getXXX_r() interfaces are mandated to take four
parameters: a pointer to the result, a pointer to an additional buffer, the
size of the buffer supplied, and a pointer to an integer to report error
values (errnop).  The returned status values are meant to be accompanied by
a corresponding error value [2] passed through the errnop pointer.

[1] https://www.gnu.org/software/libc/manual/html_node/NSS-Module-Function-Internals.html
[2] https://www.gnu.org/software/libc/manual/html_node/NSS-Modules-Interface.html

Signed-off-by: Björn Esser <besser82@fedoraproject.org>